requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pandas==2.2.3